# api/openapi.yaml
openapi: 3.1.0
info:
  title: PALISADE HTTP API
  version: 0.1.0
  description: Early behavior-first decision API. The runtime defaults to shadow mode; production decision and event calls use short-lived, session/action-bound proofs.
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
servers:
  - url: /
paths:
...
  /v1/session:
    post:
      summary: Issue a server-bound continuity cookie
      description: Backend-only endpoint. The integrating origin forwards the Set-Cookie header to the browser; cookie integrity establishes continuity, not humanity or account identity.
      operationId: issueSession
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Signed session cookie issued.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/token:
    post:
      summary: Issue a short-lived one-time proof
      operationId: issueProof
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [action]
              properties:
                session_id:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: Optional when a valid PALISADE session cookie is supplied; if both are present they must match.
                action: { type: string, minLength: 1, maxLength: 80 }
                ttl_seconds: { type: integer, minimum: 1, maximum: 300, default: 60 }
      responses:
        "201":
          description: One-time proof issued.
          content:
            application/json:
              schema:
                type: object
                required: [proof_token, expires_in]
                properties:
                  proof_token: { type: string }
                  expires_in: { type: integer }
        "401": { $ref: "#/components/responses/Error" }
  /v1/events:
    post:
      summary: Ingest privacy-limited browser events
      operationId: ingestEvents
      security:
        - eventProof: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventBatch" }
      responses:
        "202":
          description: Accepted. When event-triggered shadow evaluation is configured, the batch remains accepted even if its separate encrypted decision record is dropped.
          headers:
            X-Palisade-Shadow-Evaluation:
              description: Present only when event-triggered shadow evaluation is configured. Reports whether the accepted batch also queued a closed encrypted shadow decision.
              schema: { type: string, enum: [recorded, dropped] }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/decision:
    post:
      summary: Produce an explainable computed and enforced action decision
      operationId: decide
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecisionRequest" }
      responses:
        "200":
          description: Decision with scores and evidence.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Decision" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /v1/origin-check:
    post:
      summary: Evaluate and return the bounded origin enforcement result
      description: Alternative to /v1/decision for an origin or reverse-proxy subrequest. Production requires the one-time proof_token in the request body; signed rollouts additionally require the session cookie. A valid signed rollout may return 429 for throttle or 403 for challenge/block. Shadow and excluded canary requests return 204. Scores and evidence are not reflected in this response.
      operationId: checkOrigin
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecisionRequest" }
      responses:
        "204":
          description: Pass the request. This includes shadow mode and requests excluded from a canary.
          headers:
            X-Palisade-Decision-ID: { schema: { type: string } }
            X-Palisade-Action: { schema: { type: string, enum: [allow, observe] } }
            X-Palisade-Handling: { schema: { type: string, enum: [pass] } }
            X-Palisade-Mode: { schema: { type: string, enum: [shadow, canary, enforce] } }
            X-Palisade-Rollout-ID: { schema: { type: string } }
        "403":
          description: Apply the challenge or temporary block selected by a valid signed rollout.
          headers:
            X-Palisade-Decision-ID: { schema: { type: string } }
            X-Palisade-Action: { schema: { type: string, enum: [challenge, block] } }
            X-Palisade-Handling: { schema: { type: string, enum: [challenge, block] } }
            X-Palisade-Mode: { schema: { type: string, enum: [canary, enforce] } }
            X-Palisade-Rollout-ID: { schema: { type: string } }
            X-Palisade-Challenge-ID:
              description: Present only for challenge handling. Retrieve the bound challenge at the Location route.
              schema: { type: string, minLength: 32, maxLength: 32 }
            Location:
              description: Present only for challenge handling.
              schema: { type: string, pattern: "^/v1/challenge/[A-Za-z0-9_-]{32}$" }
            Retry-After: { schema: { type: integer, minimum: 1, maximum: 3600 } }
        "429":
          description: Apply a one-second delay response or a bounded throttle selected by a valid signed rollout. PALISADE returns immediately and never sleeps in the hot path.
          headers:
            X-Palisade-Decision-ID: { schema: { type: string } }
            X-Palisade-Action: { schema: { type: string, enum: [delay, throttle] } }
            X-Palisade-Handling: { schema: { type: string, enum: [delay, throttle] } }
            X-Palisade-Mode: { schema: { type: string, enum: [canary, enforce] } }
            X-Palisade-Rollout-ID: { schema: { type: string } }
            Retry-After: { schema: { type: integer, minimum: 1, maximum: 60 } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/challenge/{challenge_id}:
    get:
      summary: Retrieve a session-bound accessible challenge
      description: Returns a nonvisual timed-confirmation contract. It is an anti-replay step-up and does not establish that the client is human.
      operationId: getChallenge
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: challenge_id
          required: true
          schema: { type: string, minLength: 32, maxLength: 32, pattern: "^[A-Za-z0-9_-]+$" }
      responses:
        "200":
          description: Challenge metadata bound to the signed session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChallengeMetadata" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/challenge/verify:
    post:
      summary: Verify the active challenge and issue a short-lived redemption capability
      operationId: verifyChallenge
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChallengeVerifyRequest" }
      responses:
        "200":
          description: One-time redemption capability issued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChallengeVerification" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
        "425":
          description: Challenge delay has not elapsed.
          headers:
            Retry-After: { schema: { type: integer, minimum: 1, maximum: 30 } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/challenge/redeem:
    post:
      summary: Consume a challenge capability for its exact action and endpoint class
      description: Successful redemption is exactly once and authorizes only the integrating origin's bound request flow; it is recorded as challenge_passed, never human_confirmed.
      operationId: redeemChallenge
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChallengeRedeemRequest" }
      responses:
        "204":
          description: Capability consumed.
          headers:
            X-Palisade-Challenge: { schema: { type: string, enum: [redeemed] } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/challenge/fallback:
    post:
      summary: Close a challenge and request the deployment's accessible fallback
      description: PALISADE records only fallback_used; the integrating origin owns the support, reauthentication or WebAuthn route and must not include a raw return URL here.
      operationId: useChallengeFallback
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChallengeFallbackRequest" }
      responses:
        "204": { description: Fallback choice recorded and challenge closed. }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "410": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /v1/outcome:
    post:
      summary: Attach a normalized delayed outcome to a shadow session
      description: Backend-only. Outcomes are closed labels and are written only when the encrypted local shadow sink is configured.
      operationId: recordOutcome
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OutcomeRequest" }
      responses:
        "202": { description: Outcome accepted by the local shadow sink. }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminBearer:
      type: http
      scheme: bearer
      description: Distinct PALISADE_ADMIN_KEY accepted only by the loopback administrative listener.
    eventProof:
      type: apiKey
      in: header
      name: X-Palisade-Proof
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-palisade_session
  responses:
    Error:
      description: Stable machine-readable error.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string }
...
    ChallengeVerifyRequest:
      type: object
      additionalProperties: false
      required: [challenge_id, verification_token]
      properties:
        challenge_id: { type: string, minLength: 32, maxLength: 32, pattern: "^[A-Za-z0-9_-]+$" }
        verification_token: { type: string, minLength: 43, maxLength: 43, pattern: "^[A-Za-z0-9_-]+$" }
...
    ChallengeFallbackRequest:
      type: object
      additionalProperties: false
      required: [challenge_id]
      properties:
        challenge_id: { type: string, minLength: 32, maxLength: 32, pattern: "^[A-Za-z0-9_-]+$" }
    SessionResponse:
      type: object
      additionalProperties: false
      required: [session_id, expires_at]
      properties:
        session_id: { type: string, minLength: 32, maxLength: 32 }
        expires_at: { type: string, format: date-time }
...
    EventBatch:
      type: object
      additionalProperties: false
      required: [sensorVersion, events]
      properties:
        sessionId:
          type: string
          minLength: 8
          maxLength: 128
          description: Optional when a valid PALISADE session cookie is supplied; if both are present they must match.
        sensorVersion: { type: string, maxLength: 32 }
        events:
          type: array
          minItems: 1
          maxItems: 64
          items: { $ref: "#/components/schemas/BrowserEvent" }
...